legacy system: Add --enable-slurm-mpao-workaround support
Add `--enable-slurm-mpao-workaround` support so that one
can turn on the code that can work around a bug
within SLURM's `MPIR_partial_attach_ok` support.

The SLURM problem is captured in Issue #16.

In essence, `slurmstepd` uses unreliable logic to detect those
target processes that are not traced/attached by the debugger.
And as a result, `slurmstepd` may send `SIGCONT` to all
of the processes including those attached by the debugger
on one system (e.g., TOSS2) or may not send it at all (e.g., TOSS3).

This switch is added to allow the maintainer to adjust LaunchMON's behavior
accordingly. Once SLURM's issue is fixed, one will never have to
use this configuration option. But for legacy systems like
TOSS2, this option will continue to come in handy.