launchmon: Add IBM Spectrum and PowerLE support
Adjust for OpenPower ABI functional call convension
change: The ABI adds two different entry points, one
for intramodule call and the other intermodule call
(being done through TOC).

Port for IBM Spectrum jsrun. Add new spectrum definition
and also adjust back end deamon code as to how it should
synchronize itself with Spectrum MPI target.

Adjust how to handle new thread creation. When a new thread
is created, we will get the SIGTRAP notification via waitpid
on the parent thread that spawns a new thread.
But on a recent Linux kernal, we have to check the high
order bits of waitpid returned status as:

leftshift status by 8 bits and then to see if
it is equal to (SIGTRAP | (LINUX_TRACER_EVENT_CLONE << 8).

Add a work around for jsrun's broken MPIR_attach_fifo.
It is expecting ASCII 1 to be sent to the FIFO when
it should expect a numeric 1. Need to drop it when
IBM ultimately fixes the issue. (Bug filed to IBM).